Issue #1: Deadlock with (too small) buffered channel
Solved issue with a new goroutine to append the files to a slice. Uses
two barriers (sync.WaitGroup), one to wait for all search goroutines and
one to wait for the append goroutine. This way the buffered channel
won't be full if the number of files containing tag are greater than the
buffer size.
Added more comments.
Improved the structure to make it better readable.